api_yamdb/reviews/urls.py
from rest_framework import routers

from django.urls import path, include
from .views import ReviewViewSet, CommentViewSet

router = routers.DefaultRouter()
router.register(
    r'titles/(?P<title_id>\d+)/reviews',
    ReviewViewSet,
    basename='titles'
)
router.register(
    r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)',
    CommentViewSet,
    basename='comments'
)

urlpatterns = [
    path('', include(router.urls)),
]
